% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{force_adaptor}
\alias{force_adaptor}
\title{Force An Adaptor}
\usage{
force_adaptor(force_adaptor = NULL, force_cluster_type = NULL)
}
\arguments{
\item{force_adaptor}{What adaptor to force? valis values are:
"doMPI", "doParallel", "doSNOW", "doFuture", and "doMC".}

\item{force_cluster_type}{What cluster type to force?}
}
\value{

}
\description{
You can force parapurrr to ignore the "adaptor" parameters supplied in
  parapurrr function and force using a specific adaptor.
}
\examples{
\dontrun{
force_adaptor(doFuture)}
}
