% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{manual_backend}
\alias{manual_backend}
\title{Disable automatic doPar backend registering}
\usage{
manual_backend(force)
}
\arguments{
\item{force}{(logical) Stop automatic handling the doPar backend by
parapurrr? Supply TRUE to stop automatic handling of dopar backend and
force the manual mode.}
}
\value{
NULL, Internally will change the corresponding option in R
  environment.
}
\description{
By default, parapurrr automatically register doPar backends for you.
  based on the provided "adaptor" argument in the function call, parapurrr
  will automatically handle (initiate, register and terminate) the selected
  doPar backend and it's infrastructures.
  But you can disable this behaviour and force manual control of the
  doPar backend. To do that your can: \enumerate{
  \item Call manual_backend(TRUE) once in your R session. After that,
   any value for "adaptor" argument in parapurrrr function calls will be
   ignored and you will have to manually handle the doPAr backend.
  \item Call any parapurrr function with adaptor = NULL.}
}
\examples{
\dontrun{ manual_backend(TRUE) }
}
