% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{parapurrr}
\alias{parapurrr}
\alias{parapurrr-package}
\title{parapurrr: Do Purrr in Parallel}
\description{
Run purrr's mapping functions in parallel (i.e., incorporate
    multiple CPU cores instead of the default, one). The package parapurrr
    does that by connecting purrr to foreach package and its adaptors.
    Users are only required to add a prefix "pa_" before their desired
    purrr functions (e.g. pa_map instead of map). All map family functions
    and all foreach adaptors on CRAN are supported.
}
\section{Supported purrr functions}{

\enumerate{
\item \strong{map family}: map, map_chr, map_dbl, map_df, map_dfc, map_dfr, map_int, map_lgl
\item \strong{map2 family}: map2, map2_chr, map2_dbl, map2_df, map2_dfc, map2_dfr, map2_int, map2_lgl
\item \strong{conditional map family}: map_at, map_if
\item \strong{pmap family}: pmap, pmap_chr, pmap_dbl, pmap_df, pmap_dfc, pmap_dfr, pmap_int, pmap_lgl
\item \strong{imap family}: imap, imap_chr, imap_dbl, imap_df, imap_dfc, imap_dfr, imap_int, imap_lgl
\item \strong{walk family}: walk, walk2,iwalk, pwalk
}
}

\section{Supported foreach adaptors}{

\enumerate{
\item doFuture
\item doMC
\item doMPI
\item doParallel (default adaptor)
\item doRNG
\item doSNOW
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moosa-r/parapurrr}
  \item Report bugs at \url{https://github.com/moosa-r/parapurrr/issues}
}

}
\author{
\strong{Maintainer}: Moosa Rezwani \email{moosa.rezwani@gmail.com} [copyright holder]

}
\keyword{internal}
