% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{use_doRNG}
\alias{use_doRNG}
\title{Use doRNG package for reproducibility}
\usage{
use_doRNG(dorng)
}
\arguments{
\item{dorng}{(logical) Use doRNG instead of normal foreach loops?}
}
\value{
NULL, Internally will change the corresponding option in R
  environment.
}
\description{
doRNG package provides functions to perform reproducible parallel foreach
  loops. By calling this function, you can force parapurrr to use doRNG
  with combination of your any selected doPar adaptor to create fully
  reproducible parallel function calls. see:
  \href{https://cran.r-project.org/package=doRNG}{doRNG:
  Generic Reproducible Parallel Backend for 'foreach' Loops}
}
\examples{
\dontrun{
use_doRNG(TRUE)
set.seed(100)
x <- pa_map(1:3, runif)
set.seed(100)
y <- pa_map(1:3, runif)
identical(x,y)}
}
