% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{manual_register}
\alias{manual_register}
\title{Disable automatic doPar backend registering}
\usage{
manual_register(force)
}
\arguments{
\item{force}{Stop registering and handling the doPar backend by
parapurrr? Enter TRUE to stop automatic handling of dopar backends and
force the manual mode.}
}
\value{

}
\description{
By default, parapurrr automatically register doPar backends for you based
  on your provided "adaptor" argument in the function call. But you can
  disable this behaviour and force manual control of the doPar backend.
}
