% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{parapurrr}
\alias{parapurrr}
\alias{parapurrr-package}
\title{parapurrr: Do Purrr in Parallel}
\description{
This package Connects Purrr and Foreach packages. The only thing
  you need to do is to add a "pa_" suffix before your purrr functions. That's
  it! The purrr's mapping functions will be handled to Foreeach and you can
  use every available CPU cores in your computer. Of course, there are other
  extra arguments that will give you more control, but they are optional.
}
\section{Currently parapurrr implemented the parallel version of}{

\enumerate{\item map}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moosa-r/parapurrr}
  \item Report bugs at \url{https://github.com/moosa-r/parapurrr/issues}
}

}
\author{
\strong{Maintainer}: Moosa Rezwani \email{moosa.rezwani@gmail.com} [copyright holder]

}
\keyword{internal}
